// src/models/visit.model.ts
export interface Visit {
  id?: number;
  street?: string;
  city?: string;
  state?: string;
  zip?: string;
  county?: string;
  latitude?: number;
  longitude?: number;
  sqft?: number;
  condition?: string;
  crackFeet?: number;
  asphaltRepair?: number;
  drivewayUndiscounted?: number;
  drivewayQuoted?: number;
  crackUndisc?: number;
  crackQuoted?: number;
  asphalt?: number;
  totalUndisc?: number;
  totalQuoted?: number;
  totalDiscount?: number;
  firstName?: string;
  lastName?: string;
  email?: string;
  phone?: string;
  salesStatus?: string;
  followUp?: string;
  notes?: string;
  salesperson?: string;
  sprayable?: boolean;
  edgingPercent?: number;
  sessionId: string;
  hasPavers?: boolean;
  paversStairs?: boolean;
  paversWalkway?: boolean;
  paversSqft?: number;
  hasFence?: boolean;
  fenceType?: string;
  hasTrees?: boolean;
  treeAmount?: string;
  createdAt: string;
  synced: boolean;
}